Update anpr_listener.py
The anpr_listener.py script currently lacks an automatic reconnection mechanism.


The Problem: If a camera reboots or a network connection briefly drops, the SDK will trigger the disconnect_callback. Your script will log this disconnection but will 

not try to log back into the camera. This means the listener will permanently stop receiving events from that camera until you manually restart the entire container.

The Solution: We need to add a "health check" loop that periodically checks the status of each camera connection and attempts to reconnect if it has been dropped.

Here are the code modifications for anpr_listener.py to make it self-healing.